Add mod sandboxing

+++ OsiInterface/Lua/LuaBindingServer.h
@@ -345,7 +345,7 @@ namespace dse::lua
 			std::lock_guard lock(mutex_);
 
 			auto L = GetState();
-			lua_checkstack(L, args.size() + 1);
+			lua_checkstack(L, (int)args.size() + 1);
 			auto stackSize = lua_gettop(L);
 
 			try {
